Use `WriterT` instead of `StateT`
This makes the parsing code slightly "lighter" because `WriterT`
provides only the ability to append values whereas `StateT` also had the
ability to read the current value, which currently wasn't used in the
parser. Although the `ContactBuilder` creation is more complicated now
because the field that we don't care about at the moment still needs to
be initialized.